text_similarity_app/requirements.txt
streamlit
nltk
scikit-learn
PyPDF2
python-docx